Rename Sidebar in messenger to MessengerSidebar. Remove Scroll component from ChatList and ChatArea and fix styles.
