+++ src/app/[locale]/(pages)/messenger/layout.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react'
 import { getTranslations, setRequestLocale } from 'next-intl/server'
 import { AuthGuard } from '@/features/auth/ui'
 import { Typography } from '@/shared/ui'
-import { Sidebar } from '@/features/messenger/ui'
+import { MessengerSidebar } from '@/features/messenger/ui'
 import s from './MessengerLayout.module.scss'
 
 type Params = { locale: string; dialoguePartnerId?: string }
@@ -22,7 +22,7 @@ export default async function MessengerLayout({ children, params }: Props) {
     <AuthGuard requireAuth>
       <Typography variant="h1">{t('title')}</Typography>
       <div className={s.container}>
-        <Sidebar />
+        <MessengerSidebar />
         <main className={s.mainContent}>{children}</main>
       </div>
     </AuthGuard>
