+++ src/features/messenger/ui/ChatArea.tsx
@@ -1,11 +1,11 @@
 'use client'
-import { LazyLoader, Loader, Scroll } from '@/shared/ui'
+import { LazyLoader, Loader } from '@/shared/ui'
 import { User, UserCard } from '@/shared/ui/userCard'
+import { useTranslations } from 'next-intl'
 import { MessageItem } from '../api'
 import { useMessengerController } from '../hooks/useMessengerController'
 import { InputBox } from './InputBox'
 import { MessageBubble } from './MessageBubble'
-import { useTranslations } from 'next-intl'
 import s from './ChatArea.module.scss'
 
 type Props = {
@@ -29,34 +29,32 @@ export const ChatArea = ({ selectedUser, dialoguePartnerId }: Props) => {
         <UserCard user={selectedUser} />
       </header>
 
-      <Scroll className={s.scrollArea} ref={scrollRef}>
-        <div className={s.chatBody}>
-          <LazyLoader onLoadMore={loadOlderMessages} hasMore={hasMore} isFetching={isLoadingMore} />
-          {chatMessages.length ? (
-            chatMessages.map((message: MessageItem) => {
-              const { id, messageText, createdAt, status, updatedAt } = message
-              const isSender = message.ownerId !== +dialoguePartnerId
+      <div className={s.chatBody} ref={scrollRef}>
+        <LazyLoader onLoadMore={loadOlderMessages} hasMore={hasMore} isFetching={isLoadingMore} />
+        {chatMessages.length ? (
+          chatMessages.map((message: MessageItem) => {
+            const { id, messageText, createdAt, status, updatedAt } = message
+            const isSender = message.ownerId !== +dialoguePartnerId
 
-              return (
-                <MessageBubble
-                  key={id}
-                  id={id}
-                  originalMessage={messageText}
-                  isSender={isSender}
-                  userName={selectedUser.userName}
-                  avatar={selectedUser.avatar}
-                  createdAt={createdAt}
-                  updatedAt={updatedAt}
-                  status={status}
-                  onDelete={handleDelete}
-                />
-              )
-            })
-          ) : (
-            <div className={s.noMessages}>{t('noMessages')}</div>
-          )}
-        </div>
-      </Scroll>
+            return (
+              <MessageBubble
+                key={id}
+                id={id}
+                originalMessage={messageText}
+                isSender={isSender}
+                userName={selectedUser.userName}
+                avatar={selectedUser.avatar}
+                createdAt={createdAt}
+                updatedAt={updatedAt}
+                status={status}
+                onDelete={handleDelete}
+              />
+            )
+          })
+        ) : (
+          <div className={s.noMessages}>{t('noMessages')}</div>
+        )}
+      </div>
 
       <footer className={s.chatFooter}>
         <InputBox dialoguePartnerId={dialoguePartnerId} />
