+++ src/features/messenger/ui/MessengerSidebar.module.scss
@@ -1,6 +1,7 @@
+@use '@/styles/mixins' as *;
+
 .sidebar {
-  display: flex;
-  flex-direction: column;
+  @include flex-column;
   height: 80vh;
   min-width: 270px;
   border-right: 1px solid var(--color-dark-300);
@@ -18,10 +19,10 @@
 }
 
 .body {
+  @include flex-column;
   flex: 1 1 auto;
   min-height: 0;
-  display: flex;
-  flex-direction: column;
+  overflow-y: auto;
 }
 
 .usersList {
