+++ src/features/messenger/ui/MessengerSidebar.tsx
@@ -7,11 +7,11 @@ import { ChatList } from './ChatList'
 import { BaseUserList } from '@/features/userSearch/ui'
 import { useSearchUser } from '@/features/userSearch/utils/hooks'
 import { PATH } from '@/shared/config/routes'
-import s from './Sidebar.module.scss'
+import s from './MessengerSidebar.module.scss'
 import { useRouter } from '@/i18n/navigation'
 import { useLocale, useTranslations } from 'next-intl'
 
-export const Sidebar = () => {
+export const MessengerSidebar = () => {
   const t = useTranslations('messenger')
 
   const { users, isFetching, handleChangeValue, handleFetchMore, hasMore, searchValue, clearSearch } = useSearchUser({
