+++ src/features/messenger/ui/index.ts
@@ -2,4 +2,4 @@ export * from './InputBox'
 export * from './ChatList'
 export * from './ChatArea'
 export * from './MessageBubble'
-export * from './Sidebar'
+export * from './MessengerSidebar'
